refactor box plugin to remove duplicate code
