+++ src/serlo-editor/plugins/serlo-table/renderer.tsx
@@ -68,9 +68,5 @@ export function SerloTableRenderer(props: SerloTableRendererProps) {
         )}
       </Fragment>
     )
-    // Do we need to check this? Can we solve it with pure CSS?
-    // const isImage =
-    //   getDocument(content.get())(store.getState())?.plugin === 'image'
-    //   <td style={{ width: `${100 / columnHeaders.length}%` }}
   }
 }
